NEW: save and restore current date, universal resource for location and weather flow

+++ app/src/main/java/com/example/weatherapp/data/remote/dto/Current.kt
@@ -3,7 +3,7 @@ package com.example.weatherapp.data.remote.dto
 data class Current(
     val interval: Int,
     val is_day: Int,
-    val rain: Int,
+    val rain: Double,
     val temperature_2m: Double,
     val time: String,
     val wind_speed_10m: Double
